pyproject.toml
[tool.poetry]
name = "nlp-day-4"
version = "0.1.0"
description = ""
authors = ["lekien2k2 <tankien2k2@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]


[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
streamlit = "1.24.0"
pandas = "2.1.1"
numpy = "1.26.0"
scikit-learn = "1.3.1"

# Quan trọng: nâng các gói dưới đây
sentence-transformers = ">=2.6.1,<3.0.0"
datasets = ">=2.19.0,<3.0.0"
huggingface_hub = ">=0.21.2,<1.0.0"
transformers = ">=4.41.0,<5.0.0"

openai = ">=1.30,<2"
google-generativeai = ">=0.3.0"
python-dotenv = "1.0.0"
aiohttp = "<3.12"

# Nếu sau này bạn build index bằng sentence-transformers,
# nhiều máy sẽ cần torch. Có thể thêm (không bắt buộc cho prepare_data.py):
# torch = {version = ">=2.1.0", optional = true}
httpx = "<0.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

